requirements.txt
langchain==0.1.8
langchain-community==0.0.21
langchain-core==0.1.22
langchain-text-splitters==0.0.1
langchain-openai==0.0.5
pypdf==4.0.1
faiss-cpu==1.7.4
python-dotenv==1.0.1
tiktoken==0.5.2
chromadb==0.4.22
pydantic==2.6.1